--- src/components/BlueButton/index.js
@@ -1,3 +1,0 @@
-import BlueButton from './BlueButton';
-
-export default BlueButton;
