+++ src/components/ServiceMapButton/index.js
@@ -1,0 +1,3 @@
+import ServiceMapButton from './ServiceMapButton';
+
+export default ServiceMapButton;
